refactor(user): Phase 2 — email→username + Long→Integer + deletedAt 제거 (DBv5 정합)
User entity 의 email → username, deleted_at 컬럼 제거 (탈퇴 표식은
is_active=false 단일화), userId Long → Integer. UserRepository 의
findByEmail/existsByEmail → findByUsername/existsByUsername,
JpaRepository<User,Long> → <User,Integer>.

DTO 정합: UserMeResponse(username/Integer userId), SignUpRequest(username,
@Email 제거 + @Size 3~255), LoginRequest(username), AuthResponse(Integer userId).

JWT/Security 정합: SecurityUtil.currentUserIdOrNull() Long → Integer,
JwtTokenProvider.generateToken/getUserId Long → Integer, CustomUserDetailsService
의 findById Integer.parseInt, JwtAuthenticationFilter 의 userId Integer.

본 commit 단독으론 AuthService/UserMeService/SocialAuthService/controller 가
옛 시그니처(Long userId, getEmail()) 호출하고 있어 compile fail. Phase 3
(UserIdentity rename + 의존 도메인 ripple) 후 정합.

+++ src/main/java/com/naengo/api_server/global/auth/CustomUserDetailsService.java
@@ -19,7 +19,9 @@ public class CustomUserDetailsService implements UserDetailsService {
 
     @Override
     public UserDetails loadUserByUsername(String userId) throws UsernameNotFoundException {
-        User user = userRepository.findById(Long.parseLong(userId))
+        // 옵션 A — DBv5 SERIAL PK 정합으로 Integer.parseInt. Spring Security 의
+        // loadUserByUsername 인자명이 "username" 이지만 실제 값은 우리 userId 문자열.
+        User user = userRepository.findById(Integer.parseInt(userId))
                 .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
 
         if (user.isBlocked()) {
